+++ src/features/book-note/index.ts
@@ -1,1 +1,1 @@
-export { default as BookNoteForm } from './ui/book-note-form';
+export { default as BookNoteForm } from './ui/BookNoteForm/BookNoteForm';
