+++ src/features/book-note/ui/BookNoteForm/StarRating.tsx
