+++ src/features/book-note/ui/BookNoteForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from './BookNoteForm';
